Abstracting Transaction
Abstracted pretty much all the methods within Transaction, these are
quite generic and depend at most on the types of agents we have, so
methods iterating over agents by themselves (could in principle - NEED
TO RETHINK THIS - be done by implementing an iterator in environment
that would pass the list of all agents - this may be more elegant than
doing it in Transaction implementation [e.g. [banks, firms,
households]]. The above will be tested in the black_rhino_minimal and
fixed should there be any bugs.