Fixing clearing accounts :partly_sunny:
So, the way we were deleting things from accounts was not properly
working before, for some weird reason. It should work now, however a
large batch of testing will ensue. We don't actually need to delete the
transaction as it's not referenced by anything when we take it off the
books and will thus be gc'd.