Further abstractions
Added some new functions that were added during the last week or so to
abstract base classes. In particular the __getattr__ for agents and
config were added, so the design with parameters being called as
attributes despite being gathered in dictionaries for iterating. Some
methods moved to the new iterator in environment. This also simplifies
some methods which don't need to iterate over separate groups of agents
anymore (cleaner code too).