Transaction + other fixes
Removed purge accounts for singular account from Agent class, and made
it use the purge all accounts method from Transaction class. Also
removed explicit reference to deconstructor in Agent class. Added a
check in the agent lookup by ID function that raises an error when no
agent with that ID is found. In Transaction clear_accounts() have been
simplified and fixed, it now works on a reverse stack instead of for
loop (which is not a good idea when we're destroying things in the
middle of the loop), Also uses the remove_transaction() method properly
instead of deleting objects explictly. This function has been added to
avoid explicitly using deconstructor, which is bad practice. Doublec
clear_accounts have been removed. And purge_accounts has been removed
for a singular agent, as it makes no sense with the new transaction
philosophy.