Abstract rationing
Added an abstract rationing function which matches sellers and buyers
given an external priority function and an external filtering function.
If the priority function returns the same float for every input pair of
agents and filtering function returns True for every input pair of
agents this method reverts to rationing.